Change package layout
